Usuario

+++ src/main/java/uniandes/edu/co/proyecto/controller/UsuarioController.java
@@ -1,7 +1,6 @@
 package uniandes.edu.co.proyecto.controller;
 
 import uniandes.edu.co.proyecto.modelo.Usuario;
-import uniandes.edu.co.proyecto.modelo.UsuarioPK;
 import uniandes.edu.co.proyecto.repositorios.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -12,10 +11,12 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
-
 @Controller
 @RequestMapping("/usuarios")
 public class UsuarioController {
+
+    @Autowired
+    private UsuarioRepository usuarioRepository;
 
     public String index() {
         return "usuarios";
@@ -34,14 +35,14 @@ public class UsuarioController {
     }
 
     @PostMapping("rfc2Lista")
-    public String rfc2Lista(@ModelAttribute Usuario usuario,Model model) {
+    public String rfc2Lista(@ModelAttribute Usuario usuario, Model model) {
         model.addAttribute("usuarios", usuarioRepository.findUsuarioWithProductos(usuario.getTipoDeDocumento(), usuario.getNumeroDeDocumento()));
         return "rfc2Lista";
     }
 
     @PostMapping("cliente/new/save")
     public String clienteGuardar(@ModelAttribute Usuario usuario) {
-        usuarioRepository.crearUsuario(usuario.getTipoDeDocumento(), usuario.getNumeroDeDocumento(), usuario.getNombre(), usuario.getNacionalidad(), usuario.getDireccionFisica(), usuario.getCorreo(), usuario.getTelefono(), usuario.getLogin(), usuario.getPalabraClave(), usuario.getTipoPersona().name(), "cliente");
+        usuarioRepository.save(usuario);
         return "redirect:/gerente";
     }
 
@@ -50,15 +51,12 @@ public class UsuarioController {
         model.addAttribute("usuario", new Usuario());
         return "usuarioNuevo";
     }
+
     @PostMapping("usuario/new/save")
     public String usuarioGuardar(@ModelAttribute Usuario usuario) {
-        usuarioRepository.crearUsuario(usuario.getTipoDeDocumento(), usuario.getNumeroDeDocumento(), usuario.getNombre(), usuario.getNacionalidad(), usuario.getDireccionFisica(), usuario.getCorreo(), usuario.getTelefono(), usuario.getLogin(), usuario.getPalabraClave(), usuario.getTipoPersona().name(), usuario.getRol().name());
+        usuarioRepository.save(usuario);
         return "redirect:/administrador";
     }
-
-
-    @Autowired
-    private UsuarioRepository usuarioRepository;
 
     @GetMapping
     public List<Usuario> getAllUsuarios() {
@@ -68,14 +66,16 @@ public class UsuarioController {
     @GetMapping("/{tipoDeDocumento}/{numeroDeDocumento}")
     public ResponseEntity<Usuario> getUsuarioById(@PathVariable String tipoDeDocumento,
                                                   @PathVariable String numeroDeDocumento) {
-        return usuarioRepository.findById(new UsuarioPK(tipoDeDocumento, numeroDeDocumento))
-                .map(ResponseEntity::ok)
-                .orElseGet(() -> ResponseEntity.notFound().build());
+        Usuario usuario = usuarioRepository.darUsuarioPorDocumento(tipoDeDocumento, numeroDeDocumento);
+        if (usuario != null) {
+            return ResponseEntity.ok(usuario);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @PostMapping
     public ResponseEntity<Usuario> createUsuario(@RequestBody Usuario usuario) {
-        
         Usuario savedUsuario = usuarioRepository.save(usuario);
         return new ResponseEntity<>(savedUsuario, HttpStatus.CREATED);
     }
@@ -84,30 +84,32 @@ public class UsuarioController {
     public ResponseEntity<Usuario> updateUsuario(@PathVariable String tipoDeDocumento,
                                                  @PathVariable String numeroDeDocumento,
                                                  @RequestBody Usuario usuarioDetails) {
-        UsuarioPK usuarioId = new UsuarioPK(tipoDeDocumento, numeroDeDocumento);
-        return usuarioRepository.findById(usuarioId)
-                .map(usuario -> {
-                    usuario.setNombre(usuarioDetails.getNombre());
-                    usuario.setNacionalidad(usuarioDetails.getNacionalidad());
-                    usuario.setDireccionFisica(usuarioDetails.getDireccionFisica());
-                    usuario.setCorreo(usuarioDetails.getCorreo());
-                    usuario.setTelefono(usuarioDetails.getTelefono());
-                    usuario.setLogin(usuarioDetails.getLogin());
-                    usuario.setPalabraClave(usuarioDetails.getPalabraClave());
-                    usuario.setTipoPersona(usuarioDetails.getTipoPersona());
-                    usuario.setRol(usuarioDetails.getRol());
-                    return ResponseEntity.ok(usuarioRepository.save(usuario));
-                }).orElseGet(() -> ResponseEntity.notFound().build());
+        Usuario usuario = usuarioRepository.darUsuarioPorDocumento(tipoDeDocumento, numeroDeDocumento);
+        if (usuario != null) {
+            usuario.setNombre(usuarioDetails.getNombre());
+            usuario.setNacionalidad(usuarioDetails.getNacionalidad());
+            usuario.setDireccionFisica(usuarioDetails.getDireccionFisica());
+            usuario.setCorreo(usuarioDetails.getCorreo());
+            usuario.setTelefono(usuarioDetails.getTelefono());
+            usuario.setLogin(usuarioDetails.getLogin());
+            usuario.setPalabraClave(usuarioDetails.getPalabraClave());
+            usuario.setTipoPersona(usuarioDetails.getTipoPersona());
+            usuario.setRol(usuarioDetails.getRol());
+            return ResponseEntity.ok(usuarioRepository.save(usuario));
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @DeleteMapping("/{tipoDeDocumento}/{numeroDeDocumento}")
     public ResponseEntity<Void> deleteUsuario(@PathVariable String tipoDeDocumento,
                                               @PathVariable String numeroDeDocumento) {
-        UsuarioPK usuarioId = new UsuarioPK(tipoDeDocumento, numeroDeDocumento);
-        return usuarioRepository.findById(usuarioId)
-                .map(usuario -> {
-                    usuarioRepository.delete(usuario);
-                    return ResponseEntity.ok().<Void>build();
-                }).orElseGet(() -> ResponseEntity.notFound().build());
+        Usuario usuario = usuarioRepository.darUsuarioPorDocumento(tipoDeDocumento, numeroDeDocumento);
+        if (usuario != null) {
+            usuarioRepository.delete(usuario);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
